fix(ci): traffic-monitor Windows API, clippy, and workspace lint
Align traffic-monitor with windows 0.58 (BOOL, GetExtendedTcpTable), clear
vpn-engine and downstream clippy failures, and fix latent -D warnings across
the workspace so Rust and Windows cross-check jobs pass.

+++ core-service/src/metrics.rs
@@ -55,7 +55,7 @@ impl MetricsService {
             .count() as u32;
 
         Ok(MetricsSnapshot {
-            active_tunnels: self.vpn.active_count() as u32,
+            active_tunnels: self.vpn.active_count(),
             active_transports,
             blocked_requests: blocked,
             dns_queries,
